Add files via upload
This is an adventure game based on Java. There is a hunter, trapped in a randomly-generated underground maze, his goal is to find the treasure, in as few step as taken.
However, what comes to his surprise is that, once he grabbed the treasure, the underground maze is designed to collapse. Luckily, the hunter is fully aware of how many steps  he could still take before the total collasion. So he needs to get out of the cavern before the ceiling collapses. Meanwhile, since things on the ceiling are falling down,  gold bricks are dropped everywhere, so the hunter naturally wants to collect as much gold as possible along his way out.